requirements.txt
annotated-types==0.7.0
anytree==2.12.1
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
build123d==0.8.0
cachetools==5.5.1
cadquery==2.5.2
cadquery-ocp==7.7.2
casadi==3.6.7
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cqkit==0.5.8
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
Django==5.1.4
django-daisy==1.0.7
django-extensions==3.2.3
django-ninja==1.3.0
executing==2.2.0
ezdxf==1.3.5
factory_boy==3.3.1
Faker==33.1.0
Flask==3.1.0
flask-sock==0.7.0
fonttools==4.55.6
h11==0.14.0
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
multimethod==1.12
nest-asyncio==1.6.0
nlopt==2.9.1
numpy==2.2.2
ocp-tessellate==3.0.9
ocp_vscode==2.6.1
ocpsvg==0.3.4
orjson==3.10.15
packaging==24.2
parso==0.8.4
path==17.1.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
py-lib3mf==2.3.1
pyaml==25.1.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pymemcache==4.0.0
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
ruff==0.8.4
scipy==1.15.1
sentry-sdk==2.19.2
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
svgelements==1.9.6
svgpathtools==1.6.1
svgwrite==1.4.3
tornado==6.4.2
traitlets==5.14.3
trianglesolver==1.2
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.2
urllib3==2.3.0
vtk==9.3.1
wcwidth==0.2.13
webcolors==24.8.0
websockets==13.1
Werkzeug==3.1.3
wsproto==1.2.0
